created submit handler for portfolio and moved the workform into the formcontainer
