+++ src/utils/forms/submitHandlers.js
@@ -4,6 +4,51 @@ import { reset } from 'redux-form';
 
 const backendUrl = process.env.REACT_APP_BACKEND_URL;
 
+async function portfolioSubmitHandler(values, id) {
+  let { category, imageBlobs } = values;
+
+  try {
+    for (let i = 0; i < imageBlobs.length; i++) {
+      let file = imageBlobs[i];
+      let fileParts = imageBlobs[i].name.split('.');
+      let fileName = fileParts[0];
+      let fileType = fileParts[1];
+  
+      let res = await axios.post(
+        `${backendUrl}/aws-s3`,
+        {
+          fileName : fileName,
+          fileType : fileType
+        }
+      )
+      let returnData = res.data.data.returnData;
+      let signedRequest = returnData.signedRequest;
+      let signedUrl = returnData.url;
+  
+      let options = {
+        headers: {
+          'Content-Type': fileType
+        }
+      };
+
+      res = await axios.put(signedRequest, file, options)
+      
+      let params = { category, imageUrl: signedUrl };
+      
+      await axios.post(`${backendUrl}/portfolio`, params);
+    }
+
+    let res = await axios.get(`${backendUrl}/portfolio`);
+    store.dispatch({ type: 'UPDATE_PORTFOLIO_DATA', newPortfolioData: res.data });
+    store.dispatch({ type: `${category.toUpperCase()}_PORTFOLIO_DATA` });
+    store.dispatch({ type: 'DISABLE_CREATE_FORM' });
+    store.dispatch({ type: 'NOT_EDITING_FORM' });
+    store.dispatch({ type: 'SUCCESSFUL_SUBMIT' });
+  } catch (err) {
+    console.log(err);
+  }
+}
+
 async function questionsSubmitHandler(values, id) {
   const { question, answer } = values;
 
@@ -25,4 +70,7 @@ async function questionsSubmitHandler(values, id) {
   }
 }
 
-export { questionsSubmitHandler };
+export {
+  portfolioSubmitHandler,
+  questionsSubmitHandler
+};
